Increase brute speed module 22000 to ~2%

+++ OpenCL/inc_cipher_aes.cl
@@ -868,7 +868,7 @@ DECLSPEC void aes128_set_decrypt_key (PRIVATE_AS u32 *ks, PRIVATE_AS const u32 *
   aes128_InvertKey (ks, s_te1, s_td0, s_td1, s_td2, s_td3);
 }
 
-DECLSPEC void aes128_encrypt (PRIVATE_AS const u32 *ks, PRIVATE_AS const u32 *in, PRIVATE_AS u32 *out, SHM_TYPE const u32 *s_te0, SHM_TYPE const u32 *s_te1, SHM_TYPE const u32 *s_te2, SHM_TYPE const u32 *s_te3, SHM_TYPE const u32 *s_te4)
+DECLSPEC void aes128_encrypt (PRIVATE_AS const u32 *ks, PRIVATE_AS const u16 *in, PRIVATE_AS u16 *out, SHM_TYPE const u32 *s_te0, SHM_TYPE const u32 *s_te1, SHM_TYPE const u32 *s_te2, SHM_TYPE const u32 *s_te3, SHM_TYPE const u32 *s_te4)
 {
   u32 in_s[4];
 
@@ -1894,14 +1894,14 @@ DECLSPEC void AES128_set_decrypt_key (PRIVATE_AS u32 *ks, PRIVATE_AS const u32 *
 
 DECLSPEC void AES128_encrypt (PRIVATE_AS const u32 *ks, PRIVATE_AS const u32 *in, PRIVATE_AS u32 *out, SHM_TYPE u32 *s_te0, SHM_TYPE u32 *s_te1, SHM_TYPE u32 *s_te2, SHM_TYPE u32 *s_te3, SHM_TYPE u32 *s_te4)
 {
-  u32 in_s[4];
+  u16 in_s[4];
 
   in_s[0] = hc_swap32_S (in[0]);
   in_s[1] = hc_swap32_S (in[1]);
   in_s[2] = hc_swap32_S (in[2]);
   in_s[3] = hc_swap32_S (in[3]);
 
-  u32 out_s[4];
+  u16 out_s[4];
 
   aes128_encrypt (ks, in_s, out_s, s_te0, s_te1, s_te2, s_te3, s_te4);
